changing cmyk and rbg float functions to double

+++ include/dpp/utility.h
@@ -264,14 +264,14 @@ namespace dpp {
 		};
 
 		/**
-		 * @brief Convert floats to RGB for sending in embeds
+		 * @brief Convert doubles to RGB for sending in embeds
 		 * 
 		 * @param red red value, between 0 and 1 inclusive
 		 * @param green green value, between 0 and 1 inclusive
 		 * @param blue blue value, between 0 and 1 inclusive
 		 * @return uint32_t returned integer colour value
 		 */
-		uint32_t DPP_EXPORT rgb(float red, float green, float blue);
+		uint32_t DPP_EXPORT rgb(double red, double green, double blue);
 
 		/**
 		 * @brief Convert ints to RGB for sending in embeds
@@ -292,7 +292,7 @@ namespace dpp {
 		 * @param k key (black) value, between 0 and 1 inclusive
 		 * @return uint32_t returned integer colour value
 		 */
-		uint32_t DPP_EXPORT cmyk(float c, float m, float y, float k);
+		uint32_t DPP_EXPORT cmyk(double c, double m, double y, double k);
 		
 		/**
 		 * @brief Convert ints to CMYK for sending in embeds
